+++ public/index.html
@@ -35,6 +35,7 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
+    <div id="puzzle-frame"></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
