+++ src/components/PuzzleImage.html
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="./style.css"/>
+    <title>Puzzle-frame</title>
+</head>
+<body>
+    <div class='puzzle-images'>
+        <Button id="prev-btn">
+            left
+        </Button>
+        <div id='puzzle-image-1'>
+            <img class='room-img' src='./../assets/room.png' alt='an empty room with red walls'/>
+            <img id="door" src="./../assets/door.png" alt="door"/>
+            <img id="knob" src="./../assets/doorknob.png" alt="doorknob"/>
+            <img id="file-cabinet-closed" src="./../assets/filecabinetclosed.png" alt="file cabinet closed iwth a potted plant on top"/>
+        </div>
+        <div id="puzzle-image-2">
+            <img class='room-img' src='./../assets/room.png' alt='an empty room with red walls'/>
+            <img class="window" id="window-room-2" src="./../assets/window.png" alt="window seperated into four panes"/>
+            <img id="table" src="./../assets/table.png" alt="empty table"/>
+            <img id="tape-dispenser" src="./../assets/tapedispenserEmpty.png" alt="empty tape dispenser"/>
+        </div>
+        <div id="puzzle-image-3">
+            <img class='room-img' src='./../assets/room.png' alt='an empty room with red walls'/>
+            <img class="window" id="window-room-3" src="./../assets/window.png" alt="window seperated into four panes"/>
+            <img id="utility-shelf" src="./../assets/utilityShelf.png" alt="utility shelf with boxes"/>
+            <img id="target-box" src="./../assets/closedBox.png" alt="closed cardboard box"/>
+        </div>
+        <div id="puzzle-image-4">
+            <img class='room-img' src='./../assets/room.png' alt='an empty room with red walls'/>
+            <img class="window" id="window-room-4" src="./../assets/window.png" alt="window seperated into four panes"/>
+            <img id="shredder" src="./../assets/shredder.png" alt="shredder"/>
+        </div>
+        
+<!-- conditional rendering of items -->
+        <Button id="next-btn">
+            right
+        </Button>
+    </div>
+</body>
+</html>
